+++ database/migrations/2021_10_02_091738_create_conclusions_table.php
@@ -28,7 +28,7 @@ class CreateConclusionsTable extends Migration
             $table->text('audit_report');
             $table->text('audited_financial_statement');
             $table->text('other_financial_info');
-            $table->integer('status')->default(0);
+            $table->unsignedSmallInteger('status')->default(0);
             $table->timestamps();
         });
     }
