+++ database/migrations/2021_10_05_075845_create_engagement_invites_table.php
@@ -23,7 +23,7 @@ class CreateEngagementInvitesTable extends Migration
             $table->foreign('engagement_team_role_id')->references('id')->on('engagement_team_roles');
             $table->unsignedInteger('company_id');
             $table->foreign('company_id')->references('id')->on('companies');
-            $table->integer('status')->default(0);
+            $table->unsignedSmallInteger('status')->default(0);
             $table->softDeletes();
             $table->timestamps();
         });
